requirements.txt
folium==0.14.0
lxml==4.9.3
numpy==1.24.3
pandas==2.0.3
Pillow==9.4.0
Pillow==10.1.0
Requests==2.31.0
streamlit==1.27.2
streamlit_folium==0.15.0
